AbstractJsonRpcAuthenticationAcceptanceTest
